refactor(discord): extract empty response guidance
