+++ src/services/discord/turn_bridge/terminal_outcome_delivery/empty_response_recovery.rs
@@ -1,16 +1,6 @@
 //! Empty-response recovery and silent-turn terminal delivery for terminal outcome delivery.
-
-use std::sync::Arc;
-
-use super::*;
-
-fn empty_response_guidance(rx_disconnected: bool) -> &'static str {
-    if rx_disconnected {
-        "⚠️ provider 프로세스가 응답을 만들기 전에 종료됐어요.\n자동 복구에서도 전달할 응답을 찾지 못했습니다. 같은 요청을 다시 시도해 주세요. 문제가 반복되면 `!clear`로 세션을 초기화한 뒤 다시 보내 주세요."
-    } else {
-        "⚠️ provider가 응답 내용 없이 턴을 종료했어요.\n자동 복구에서도 전달할 응답을 찾지 못했습니다. 같은 요청을 다시 시도해 주세요. 문제가 반복되면 `!clear`로 세션을 초기화한 뒤 다시 보내 주세요."
-    }
-}
+mod guidance;
+use {super::*, std::sync::Arc};
 
 pub(super) enum EmptyResponseRecoveryMessage {
     ResumeFailureAlreadyHandled,
@@ -272,7 +262,7 @@ pub(super) async fn handle_empty_response_recovery(
                 if !resume_failed {
                     if rx_disconnected {
                         terminal_empty_response_notice =
-                            Some(empty_response_guidance(true).to_string());
+                            Some(guidance::empty_response_guidance(true).to_string());
                         let ts = chrono::Local::now().format("%H:%M:%S");
                         tracing::warn!(
                             "  [{ts}] ⚠ Empty response: rx disconnected before any text \
@@ -283,7 +273,7 @@ pub(super) async fn handle_empty_response_recovery(
                         );
                     } else {
                         terminal_empty_response_notice =
-                            Some(empty_response_guidance(false).to_string());
+                            Some(guidance::empty_response_guidance(false).to_string());
                         let ts = chrono::Local::now().format("%H:%M:%S");
                         tracing::warn!(
                             "  [{ts}] ⚠ Empty response: done without text (channel {})",
@@ -424,32 +414,5 @@ pub(super) async fn handle_empty_response_recovery(
             spoken_delivery_response,
             resume_retry_queued,
         }
-    }
-}
-
-#[cfg(test)]
-mod guidance_message_tests {
-    use super::empty_response_guidance;
-
-    #[test]
-    fn disconnected_empty_response_explains_recovery_and_next_actions() {
-        let response = empty_response_guidance(true);
-
-        assert!(response.contains("프로세스가 응답을 만들기 전에 종료"));
-        assert!(response.contains("자동 복구"));
-        assert!(response.contains("다시 시도"));
-        assert!(response.contains("`!clear`"));
-        assert!(!response.contains("(No response"));
-    }
-
-    #[test]
-    fn completed_empty_response_explains_recovery_and_next_actions() {
-        let response = empty_response_guidance(false);
-
-        assert!(response.contains("응답 내용 없이 턴을 종료"));
-        assert!(response.contains("자동 복구"));
-        assert!(response.contains("다시 시도"));
-        assert!(response.contains("`!clear`"));
-        assert!(!response.contains("(No response"));
     }
 }
